requirements.txt
streamlit
requests
boto3
apache-airflow
python-dotenv
tweepy
facebook-sdk
google-api-python-client
praw
